adding offset toggle capability
Two of the digital pins are used to offset the signal so negative voltages can be read. Not yet implemented on STM32 but it will work the same way. Some crosstalk between channels evident on Arduino Uno using LM324N. Substantial with our usual prescaler of 8, only faintly visible when using a prescaler of 16.